Improves readability of logs.

+++ tests/scripts/free-unused-resources.sh
@@ -1,7 +1,15 @@
 #!/usr/bin/env bash
 
+echo " "
 echo "Freeing unused resources... "
 
-docker ps -a | grep 'Exited .* ago' | sed 's/\s\+/ /g' | cut -d ' ' -f 1 | while read id; do echo "docker rm $id"; docker rm $id; done
+if docker ps -a | grep -q 'Exited .* ago' ; then
+    docker ps -a | grep 'Exited .* ago' | sed 's/\s\+/ /g' | cut -d ' ' -f 1 | while read id; do echo "docker rm $id"; docker rm $id; done
+else
+    echo "No docker containers to remove. "
+fi
 
-oc adm prune images --keep-tag-revisions=1 --keep-younger-than=30m --confirm
+oc adm prune images --keep-tag-revisions=1 --keep-younger-than=30m --confirm || true
+
+echo " "
+exit 0
